pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-development"
version = "0.0.1"
description = "LangChain Agent"
authors = [{ name = "yongzhiyin", email = "610415432@qq.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=1.0.0",
    "langchain>=1.1.0",
    "langchain-core>=1.0.0",  # 之前建议放宽的约束
    "python-dotenv>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",     # 新增PostgreSQL
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

# ---- 关键：标准 src-layout 打包配置 ----
[tool.setuptools]
package-dir = { "" = "src" } # 路径寻找

[tool.setuptools.packages.find]
where = ["src"]
include = ["agent*", "chat_models*", "utils*"] # 对应的包名

# ----------------------------------------
